[REF] pms_api_rest: consistency in datamodels and services filenames
