+++ pms_api_rest/datamodels/__init__.py
@@ -1,28 +1,20 @@
-from . import pms_calendar_search_param
-from . import pms_calendar_info
+from . import pms_calendar
 
-from . import pms_folio_search_param
-from . import pms_folio_info
+from . import pms_folio
 
-from . import pms_room_info
-from . import pms_room_type_info
-from . import pms_room_search_param
+from . import pms_room
+from . import pms_room_type
 
-from . import pms_reservation_info
+from . import pms_reservation
 
-from . import pms_checkin_partner_info
-from . import pms_partner_info
+from . import pms_checkin_partner
+from . import pms_partner
 
-from . import pms_calendar_swap_info
-from . import pms_calendar_changes
+from . import pms_property
+from . import pms_account_journal
+from . import pms_payment
 
-from . import pms_property_info
-from . import pms_property_search_param
-from . import pms_account_journal_info
-from . import pms_payment_info
-from . import user_input
-from . import user_output
+from . import pms_user
 
-from . import pms_pricelist_info
-from . import pms_pricelist_item_search_param
-from . import pms_pricelist_item_info
+from . import pms_pricelist
+from . import pms_pricelist_item
