+++ pms_api_rest/datamodels/pms_calendar.py
@@ -3,6 +3,25 @@ from marshmallow import fields
 from odoo.addons.datamodel.core import Datamodel
 
 
+class PmsCalendarChanges(Datamodel):
+    _name = "pms.calendar.changes"
+    reservationLinesChanges = fields.List(fields.Dict(required=False, allow_none=True))
+
+
+class PmsCalendarSwapInfo(Datamodel):
+    _name = "pms.calendar.swap.info"
+    swapFrom = fields.String(required=True, allow_none=False)
+    swapTo = fields.String(required=True, allow_none=False)
+    roomIdA = fields.Integer(required=True, allow_none=False)
+    roomIdB = fields.Integer(required=True, allow_none=False)
+
+
+class PmsCalendarSearchParam(Datamodel):
+    _name = "pms.calendar.search.param"
+    date_from = fields.String(required=False, allow_none=True)
+    date_to = fields.String(required=False, allow_none=True)
+
+
 class PmsCalendarInfo(Datamodel):
     _name = "pms.calendar.info"
     id = fields.Integer(required=False, allow_none=True)
