+++ pms_api_rest/datamodels/pms_checkin_partner.py
