+++ pms_api_rest/datamodels/pms_folio.py
@@ -3,6 +3,13 @@ from marshmallow import fields
 from odoo.addons.datamodel.core import Datamodel
 
 
+class PmsFolioSearchParam(Datamodel):
+    _name = "pms.folio.search.param"
+
+    date_from = fields.String(required=False, allow_none=True)
+    date_to = fields.String(required=False, allow_none=True)
+
+
 class PmsFolioInfo(Datamodel):
     _name = "pms.folio.info"
     id = fields.Integer(required=False, allow_none=True)
