--- pms_api_rest/datamodels/pms_folio_search_param.py
@@ -1,10 +1,0 @@
-from marshmallow import fields
-
-from odoo.addons.datamodel.core import Datamodel
-
-
-class PmsFolioSearchParam(Datamodel):
-    _name = "pms.folio.search.param"
-
-    date_from = fields.String(required=False, allow_none=True)
-    date_to = fields.String(required=False, allow_none=True)
