--- pms_api_rest/datamodels/pms_room_info.py
@@ -1,11 +1,0 @@
-from marshmallow import fields
-
-from odoo.addons.datamodel.core import Datamodel
-
-
-class PmsRoomInfo(Datamodel):
-    _name = "pms.room.info"
-    id = fields.Integer(required=False, allow_none=True)
-    name = fields.String(required=False, allow_none=True)
-    roomTypeId = fields.Integer(required=False, allow_none=True)
-    capacity = fields.Integer(required=False, allow_none=True)
