--- pms_api_rest/datamodels/pms_room_type_info.py
@@ -1,9 +1,0 @@
-from marshmallow import fields
-
-from odoo.addons.datamodel.core import Datamodel
-
-
-class PmsRoomTypeInfo(Datamodel):
-    _name = "pms.room.type.info"
-    id = fields.Integer(required=False, allow_none=True)
-    name = fields.String(required=False, allow_none=True)
