+++ pms_api_rest/datamodels/pms_user.py
@@ -3,6 +3,12 @@ from marshmallow import fields
 from odoo.addons.datamodel.core import Datamodel
 
 
+class PmsApiRestUserInput(Datamodel):
+    _name = "pms.api.rest.user.input"
+    username = fields.String(required=False, allow_none=True)
+    password = fields.String(required=False, allow_none=True)
+
+
 class PmsApiRestUserOutput(Datamodel):
     _name = "pms.api.rest.user.output"
     # user = fields.String(required=False, allow_none=True)
