+++ pms_api_rest/services/__init__.py
@@ -1,10 +1,10 @@
-from . import folio_services
-from . import room_services
-from . import room_type_services
-from . import calendar_service
-from . import partner_services
+from . import pms_folio_service
+from . import pms_room_service
+from . import pms_room_type_services
+from . import pms_calendar_service
+from . import pms_partner_service
 
-from . import reservation_services
-from . import property_services
-from . import login_service
-from . import pricelist_service
+from . import pms_reservation_service
+from . import pms_property_service
+from . import pms_login_service
+from . import pms_pricelist_service
