+++ pms_api_rest/services/pms_login_service.py
@@ -10,7 +10,7 @@ from odoo.addons.base_rest_datamodel.restapi import Datamodel
 from odoo.addons.component.core import Component
 
 
-class PmsPartnerService(Component):
+class PmsLoginService(Component):
     _inherit = "base.rest.service"
     _name = "pms.auth.service"
     _usage = "login"
